task 3 fix--please work

+++ sort.h
@@ -26,7 +26,7 @@ void bubble_sort(int *array, size_t size);
 void insertion_sort_list(listint_t **list);
 void selection_sort(int *array, size_t size);
 void quick_sort(int *array, size_t size);
-int arr_partition(int *array, int low_idx, int  high_idx, int len);
-void quick_sort_recursive(int *arr, int low_idx, int high_idx, int len);
+int arr_partition(int *array, int low_idx, int  high_idx, size_t len);
+void quick_sort_recursive(int *arr, int low_idx, int high_idx, size_t len);
 
 #endif
